feat: replace burned WBNB button with Claim rewards button
