+++ .changeset/wet-signs-call.md
@@ -1,0 +1,5 @@
+---
+"@venusprotocol/evm": minor
+---
+
+update navbar
