+++ apps/evm/src/components/Icon/icons/index.ts
@@ -79,3 +79,4 @@ export { default as chevronDown } from './chevronDown';
 export { default as user } from './user';
 export { default as expand } from './expand';
 export { default as dollar } from './dollar';
+export { default as parachute } from './parachute';
