+++ apps/evm/src/containers/Layout/NavBar/ClaimRewardsButton/__testUtils__/fakeData.ts
