+++ apps/evm/src/containers/Layout/NavBar/ClaimRewardsButton/types.ts
