+++ apps/evm/src/pages/Markets/Assets/index.tsx
@@ -30,6 +30,7 @@ export const Assets: React.FC<AssetsProps> = ({ pool }) => (
         orderDirection: 'desc',
       }}
       header={<AdBanner />}
+      className="border-0 lg:border"
     />
   </div>
 );
